refactor: delete conftest from src/libecalc/fixtures
