+++ src/libecalc/fixtures/__init__.py
@@ -3,9 +3,3 @@ from .cases.all_energy_usage_models import *  # noqa: F403
 from .cases.compressor_systems_and_compressor_train_temporal import compressor_systems_and_compressor_train_temporal
 from .cases.ltp_export import ltp_export_yaml
 from .compressor_process_simulations.compressor_process_simulations import *  # noqa: F403
-from .conftest import (
-    fuel_gas,
-    medium_fluid_dto,
-    predefined_variable_speed_compressor_chart_dto,
-    rich_fluid_dto,
-)
