+++ tests/libecalc/core/conftest.py
@@ -8,6 +8,7 @@ import libecalc.common.energy_usage_type
 import libecalc.common.fixed_speed_pressure_control
 import libecalc.common.serializable_chart
 import libecalc.dto.fuel_type
+from libecalc.common.serializable_chart import ChartCurveDTO, VariableSpeedChartDTO
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated import TabularConsumerFunction
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated.common import VariableExpression
 from libecalc.domain.infrastructure.energy_components.turbine import Turbine
@@ -24,7 +25,7 @@ from libecalc.testing.yaml_builder import YamlTurbineBuilder
 
 
 @pytest.fixture
-def medium_fluid() -> FluidModel:
+def fluid_model_medium() -> FluidModel:
     return FluidModel(
         eos_model=EoSModel.SRK,
         composition=FluidComposition.model_validate(MEDIUM_MW_19P4),
@@ -32,7 +33,7 @@ def medium_fluid() -> FluidModel:
 
 
 @pytest.fixture
-def rich_fluid() -> FluidModel:
+def fluid_model_rich() -> FluidModel:
     return FluidModel(
         eos_model=EoSModel.SRK,
         composition=FluidComposition.model_validate(RICH_MW_21P4),
@@ -40,7 +41,7 @@ def rich_fluid() -> FluidModel:
 
 
 @pytest.fixture
-def dry_fluid() -> FluidModel:
+def fluid_model_dry() -> FluidModel:
     return FluidModel(
         eos_model=EoSModel.SRK,
         composition=FluidComposition.model_validate(DRY_MW_18P3),
@@ -239,10 +240,10 @@ def variable_speed_compressor_chart_dto() -> libecalc.common.serializable_chart.
 
 @pytest.fixture
 def variable_speed_compressor_train_dto(
-    medium_fluid_dto, variable_speed_compressor_chart_dto
+    fluid_model_medium, variable_speed_compressor_chart_dto
 ) -> dto.VariableSpeedCompressorTrain:
     return dto.VariableSpeedCompressorTrain(
-        fluid_model=medium_fluid_dto,
+        fluid_model=fluid_model_medium,
         stages=[
             dto.CompressorStage(
                 compressor_chart=variable_speed_compressor_chart_dto,
@@ -260,10 +261,10 @@ def variable_speed_compressor_train_dto(
 
 @pytest.fixture
 def variable_speed_compressor_train_two_stages_dto(
-    medium_fluid_dto, variable_speed_compressor_chart_dto
+    fluid_model_medium, variable_speed_compressor_chart_dto
 ) -> dto.VariableSpeedCompressorTrain:
     return dto.VariableSpeedCompressorTrain(
-        fluid_model=medium_fluid_dto,
+        fluid_model=fluid_model_medium,
         stages=[
             dto.CompressorStage(
                 compressor_chart=variable_speed_compressor_chart_dto,
@@ -353,3 +354,92 @@ def tabular_consumer_function_factory():
         )
 
     return create_tabular_consumer_function
+
+
+@pytest.fixture
+def predefined_variable_speed_compressor_chart_dto() -> VariableSpeedChartDTO:
+    """NOT USED CURRENTLY, KEPT FOR FUTURE REFERENCE"""
+    return VariableSpeedChartDTO(
+        curves=[
+            ChartCurveDTO(
+                speed_rpm=7689.0,
+                rate_actual_m3_hour=[2900.0666, 3503.8068, 4002.5554, 4595.0148],
+                polytropic_head_joule_per_kg=[82530.702036, 78443.25272100001, 72239.03594100001, 60107.539661999996],
+                efficiency_fraction=[0.723, 0.7469, 0.7449, 0.7015],
+            ),
+            ChartCurveDTO(
+                speed_rpm=8787.0,
+                rate_actual_m3_hour=[3305.5723, 4000.1546, 4499.2342, 4996.8728, 5241.9892],
+                polytropic_head_joule_per_kg=[
+                    107428.875417,
+                    101959.123527,
+                    95230.48671000001,
+                    84305.752866,
+                    78230.827962,
+                ],
+                efficiency_fraction=[0.7241, 0.7449, 0.7464, 0.722, 0.7007],
+            ),
+            ChartCurveDTO(
+                speed_rpm=9886.0,
+                rate_actual_m3_hour=[3708.8713, 4502.2531, 4993.5959, 5507.8114, 5924.3308],
+                polytropic_head_joule_per_kg=[
+                    135823.185648,
+                    129322.210482,
+                    121892.878719,
+                    110621.46830400001,
+                    98633.21175900001,
+                ],
+                efficiency_fraction=[0.723, 0.7473, 0.748, 0.7306, 0.704],
+            ),
+            ChartCurveDTO(
+                speed_rpm=10435.0,
+                rate_actual_m3_hour=[3928.0389, 4507.4654, 5002.1249, 5498.9912, 6248.5937],
+                polytropic_head_joule_per_kg=[
+                    151422.09804,
+                    146980.631331,
+                    140775.67978200002,
+                    131074.593345,
+                    109705.500756,
+                ],
+                efficiency_fraction=[0.7232, 0.7437, 0.7453, 0.7414, 0.701],
+            ),
+            ChartCurveDTO(
+                speed_rpm=10984.0,
+                rate_actual_m3_hour=[4138.6974, 5002.4758, 5494.3704, 6008.6962, 6560.148],
+                polytropic_head_joule_per_kg=[
+                    167543.961912,
+                    159657.01326900002,
+                    151353.646803,
+                    139907.430036,
+                    121474.81477800001,
+                ],
+                efficiency_fraction=[0.7226, 0.7462, 0.7468, 0.7349, 0.7023],
+            ),
+            ChartCurveDTO(
+                speed_rpm=11533.0,
+                rate_actual_m3_hour=[4327.9175, 4998.517, 5505.8851, 6027.6167, 6506.9064, 6908.2832],
+                polytropic_head_joule_per_kg=[
+                    185235.416955,
+                    178887.22567200003,
+                    171985.250079,
+                    161764.148295,
+                    147514.415994,
+                    133600.348539,
+                ],
+                efficiency_fraction=[0.7254, 0.7444, 0.745, 0.7466, 0.7266, 0.7019],
+            ),
+            ChartCurveDTO(
+                speed_rpm=10767.0,
+                rate_actual_m3_hour=[4052.9057, 4500.6637, 4999.41, 5492.822, 6000.6263, 6439.4876],
+                polytropic_head_joule_per_kg=[
+                    161345.07,
+                    157754.61000000002,
+                    152506.26,
+                    143618.4,
+                    131983.74000000002,
+                    117455.13,
+                ],
+                efficiency_fraction=[0.724, 0.738, 0.7479, 0.74766, 0.7298, 0.7014],
+            ),
+        ],
+    )
